+++ changelog.md
@@ -10,6 +10,9 @@ To create a recipe, the `-r <recipe_name>` can be used. To list available
 recipes, the `--recipe-list` and `--recipe-list-short` options were added. 
 - Added `-ft` or `--fetch-tags` which allows to retrieve all DockerHub 
 container tags.
+- Added function to collect all the components from the components classes,
+replacing the current process_map dictionary implementation. Now, it will be
+generated from the engine rather than hardcoded into the dict.
 
 ### Components changes